refactored old code added design docs updated versions for mastermind and other dependencies

+++ train.py
@@ -5,5 +5,6 @@ nltk.download('wordnet')
 
 train_path = os.path.join("data", "models", "general_model")
 dataset = os.path.join("data", "general_dataset.json")
-shutil.rmtree(train_path)
+if(os.path.exists(train_path)):
+  shutil.rmtree(train_path)
 mastermind.generate_model(dataset, train_path, "Hi")
